Rename src/Router.js to src/routes.js
